Fix template filter setup
