+++ council_finance/templates/council_finance/my_lists.html
@@ -1,4 +1,5 @@
 {% extends "base.html" %}
+{% load extras %}
 {% block title %}My Lists - Council Finance Counters{% endblock %}
 {% block content %}
 <h1 class="text-2xl font-bold mb-4">My Lists</h1>
