+++ council_finance/templatetags/__init__.py
@@ -1,10 +1,1 @@
-from django import template
-
-register = template.Library()
-
-@register.filter
-def get_item(obj, key):
-    """Retrieve dictionary value by key in templates."""
-    if isinstance(obj, dict):
-        return obj.get(key)
-    return None
+# Register template tags package
